созданы файлы ресурсов

+++ src/main/java/org/aleh4min/labtask1/dto/user/UserRequestDto.java
@@ -10,7 +10,7 @@ public class UserRequestDto {
     private Long id;
     private String firstName;
     private String lastName;
-    private byte age;
+    private short age;
     private String email;
     private List<AddressRequestDto> addresses;
 }
